requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-bootstrap3==9.0.0
django-db-file-storage==0.4.2
django-filter==1.0.4
django-forms-bootstrap==3.1.0
django-salesforce==0.7.2
django-tables2==1.11.0
djangorestframework==3.6.4
et-xmlfile==1.0.1
idna==2.6
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.5
openpyxl==2.4.8
pytz==2017.2
PyYAML==3.12
requests==2.18.4
simplejson==3.11.1
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
